requirements.txt
pillow
thefuzz
python-Levenshtein
pyinstaller
